python/requirements.txt
Flask==3.1.1
bcrypt
PyJWT
python-dotenv